assignment-solver/backend/python_requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
PyPDF2
reportlab
python-dotenv
requests
pypdf
langchain-google-genai
